+++ Apps/Examples/Examples/All Examples/FitCameraToGeometryExample.swift
@@ -1,9 +1,7 @@
 import UIKit
 import MapboxMaps
-import Turf
 
 @objc(FitCameraToGeometryExample)
-
 public class FitCameraToGeometryExample: UIViewController, ExampleProtocol {
 
     internal var mapView: MapView!