+++ Apps/Examples/Examples/All Examples/SwiftUIExample.swift
@@ -1,6 +1,5 @@
 import UIKit
 import MapboxMaps
-import Turf
 import SwiftUI
 
 internal struct Camera {