+++ Sources/MapboxMaps/Foundation/Camera/CameraAnimationsManager.swift
@@ -1,5 +1,4 @@
 import UIKit
-import Turf
 @_implementationOnly import MapboxCommon_Private
 
 public protocol CameraAnimator: Cancelable {