+++ Sources/MapboxMaps/Foundation/Extensions/Turf/Geometry.swift
@@ -1,4 +1,3 @@
-import Turf
 import MapboxCommon
 
 // swiftlint:disable cyclomatic_complexity