+++ Sources/MapboxMaps/Offline/TileStore+MapboxMaps.swift
@@ -155,9 +155,9 @@ extension TileStore {
     ///     worker thread; it is the responsibility of the user to dispatch to a
     ///     user-controlled thread.
     public func tileRegionGeometry(forId id: String,
-                                   completion: @escaping (Result<Geometry, Error>) -> Void) {
+                                   completion: @escaping (Result<MapboxCommon.Geometry, Error>) -> Void) {
         __getTileRegionGeometry(forId: id,
-                                callback: tileStoreClosureAdapter(for: completion, type: Geometry.self))
+                                callback: tileStoreClosureAdapter(for: completion, type: MapboxCommon.Geometry.self))
     }
 
     /// Fetch a tile region's associated metadata