+++ Sources/MapboxMaps/Style/Style.swift
@@ -1,4 +1,3 @@
-import Turf
 @_implementationOnly import MapboxCommon_Private
 @_implementationOnly import MapboxCoreMaps_Private