+++ Tests/MapboxMapsTests/Foundation/GeoJSON/MultiLineStringTests.swift
@@ -1,6 +1,5 @@
 import XCTest
 import CoreLocation
-import Turf
 @testable import MapboxMaps
 
 // Disabling rules against force try for this test file.