+++ Tests/MapboxMapsTests/Snapshot/MapboxMapsSnapshotTests.swift
@@ -1,6 +1,5 @@
 import XCTest
 import MapboxMaps
-import Turf
 import Foundation
 import UIKit