+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
+### Features ✨ and improvements 🏁
+
+* It is no longer necessary to `import Turf`. [#622](https://github.com/mapbox/mapbox-maps-ios/pull/622)
+* `Geometry` now refers to `Turf.Geometry` instead of `MapboxCommon.Geometry`. [#622](https://github.com/mapbox/mapbox-maps-ios/pull/622)
+
 ## 10.0.0-rc.7 - August 25, 2021
 
 ### Features ✨ and improvements 🏁